requirements.txt
# Python
openai>=1.99.0
azure-ai-contentsafety==1.0.0
azure-identity==1.18.0
azure-appconfiguration==1.7.2
azure-servicebus==7.12.3
msal==1.31.0

# Agent Framework (requires --pre flag)
agent-framework-azure-ai

# HTTP Requests (for Microsoft Graph API REST calls)
requests==2.32.5
flask==3.1.0

# Email notifications
azure-communication-email==1.0.0

# Configuration & Environment
python-dotenv==1.0.1
pydantic>=2.11.0
pydantic-settings==2.7.0

# Utilities
aiohttp==3.11.11
tenacity==9.0.0

# Logging & Monitoring
structlog==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-mock==3.14.0

# UI
streamlit==1.41.0
